requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
python-dotenv==1.0.1
crewai==0.41.0
crewai-tools==0.15.0
google-generativeai==0.7.2
pydantic==2.6.3
httpx==0.27.0
